eml export WIP
